#59 clear test-only export: extract seed catalogue, drop listPrompts
knip's production pass flagged `listPrompts` in src/prompts/registry.ts as
unused-in-production: only registry.test.ts and coverage.test.ts walked the
full registry; production uses point lookups via lookupPrompt. Extracted
the seed catalogue (RuleSeed, supplementalSeeds, buildPrompt, and a new
allSeeds()) into src/prompts/seeds.ts, which buildRegistry now consumes in
production. Deleted listPrompts; tests enumerate via allSeeds() and add a
check that every seed is reachable through the production lookupPrompt path.
allSeeds() dedups last-wins to match the registry Map, so severities are
unchanged.
